+++ src/main/scala/xitrum/validator/Email.scala
@@ -1,8 +1,13 @@
 package xitrum.validator
 
-object Email extends Validator {
-  def v(name: String, value: Any) =
-    if ("""(?i)^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$""".r.findFirstIn(value.asInstanceOf[String]).isDefined)
+object Email extends Validator[String] {
+  private val PATTERN = """(?i)^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$""".r
+
+  def check(value: String) =
+    PATTERN.findFirstIn(value).isDefined
+
+  def message(name: String, value: String) =
+    if (check(value))
       None
     else
       Some("%s must be an email address".format(name))