+++ src/main/scala/xitrum/validator/MinLength.scala
@@ -4,9 +4,11 @@ object MinLength {
   def apply(length: Int) = new MinLength(length)
 }
 
-class MinLength(min: Int) extends Validator {
-  def v(name: String, value: Any) =
-    if (value.asInstanceOf[String].length >= min)
+class MinLength(min: Int) extends Validator[String] {
+  def check(value: String) = value.length >= min
+
+  def message(name: String, value: String) =
+    if (value.length >= min)
       None
     else
       Some("%s must be at least %d characters".format(name, min))