+++ src/main/scala/xitrum/validator/Required.scala
@@ -1,8 +1,10 @@
 package xitrum.validator
 
-object Required extends Validator {
-  def v(name: String, value: Any) =
-    if (value.asInstanceOf[String].trim.isEmpty)
+object Required extends Validator[String] {
+  def check(value: String) = !value.trim.isEmpty
+
+  def message(name: String, value: String) =
+    if (value.trim.isEmpty)
       Some("%s must not be empty".format(name))
     else
       None